backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
DateTime==5.5
deprecation==2.1.0
filelock==3.16.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
google==3.0.0
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
nltk==3.9.1
packaging==24.1
postgrest==0.16.11
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.0.2
regex==2024.9.11
setuptools==74.1.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.13.2
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
uuid==1.30
websockets==12.0
Werkzeug==3.0.4
yarl==1.11.1
zope.interface==7.0.3
cachetools==5.5.0
cffi==1.17.0
charset-normalizer==3.3.2
distlib==0.3.8
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
huggingface-hub==0.24.7
lxml==5.3.0
ntscraper==0.3.17
numpy==2.1.1
oauthlib==3.2.2
outcome==1.3.0.post0
pillow==10.2.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.5
selenium==4.23.1
setuptools==70.0.0
sortedcontainers==2.4.0
transformers==4.44.2
tweepy==4.14.0
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
wsproto==1.2.0